requirements.txt
joblib==1.5.3
numpy==1.26.4
pillow==12.0.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
tensorboard==2.20.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu113
--find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.11.0+cu113
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.12.0+cu113
